Update Dockerfile to optimise build time
Having the test dependencies install in the final build stage prevents
effective caching of the layers, because the application code it in a
previous stage, each change means we have to run a fresh `test` stage
every time and it is the one of the longest running stages.

Instead we install them early, but only if the RAILS_ENV is set to
`test`.

We've also update the Docker file, Just fixing the glaring errors here
that get raised in Github.