Incorporate shellcheck into job
For some reason when we run shellcheck locally it fails, but in CI it
appears to pass. This work fixes the errors.

The existing CI action runs shellcheck via a Rails task, the reason
seems to be that we have included some Ruby scripts in the script
directory and the task excludes those.

We have changed this by moving the Ruby scripts, which we are not sure
are needed any longer and getting rid of the Rails tasks. This means we
can run shellcheck without requiring the Rails environment to be setup.

Because these untested Ruby scripts are now in the `app` directory, we
have to exclude them from the test coverage report.

With that work done, we can run chellcheck from a simple `RUN` command
in the action.