Use Rails 6.1 defaults
We've been running Rails 6.1 for a while now and inline with the upgrade
documentation, we should now be using the 6.1 defaults:

https://guides.rubyonrails.org/upgrading_ruby_on_rails.html#configure-framework-defaults

Rather than change each individual setting, we've gone all in and taken
confidence from our extensive test suite that everything is good.

This clears a path for the Rails 6 to 7 upgrade.