Fix org_participation role
This enum stores the values as integers and maps them to symbols for
ease. We have also declared is as a string, which is both incorrect and
broken in Rails 7.

In Rails 7 we can declare a default inside the enum definition, so we do
that.

We think this is working now because all records are 'implementing/3' so
any loaded model returns the string `implementing` by default.

We've check the production database and all values are correctly set to
the integer '3'.