Allow Users to be anonymised
We want to anonymise users instead of deleting them, this is less work
to implement and retains some history for DSIT whilst meeting their GDPR
requirements.

The user life cycle will be:

Active > Deactivated > Anonymised

Before we can anonymise users, we have to know how long they have been
deactivated for as we want to retain users for five years before
anonymising them.

We will use DateTime to store when a user is deactivated, this will
replace the `active` attribute:

- when the deactivated_at is nil the user is active
- when the deactivated_at has a date time the user has been deactivated

We have no way to know when a user was deactivated (hence the work!) so
we'll use the updated_at to be our best guess for date migration.

We create an `active` method on User and a `active?` alias, to replicate
the old behaviour and update the two scopes so they work based on
deactivated_at.

We adjust the specs setup to make sure we are setting a
deactivated_at date not the old active value.

Finally we enable users to be deactivated and re-activated in the
same was as they were before, this work is temporary as we plan to
introduce a new UI for managing the state of a user.

We update the service that updates users so it can set or nil the
deactivated_at date and time, we must set it to nil when re-activating.