Migration to allow users to be anonymised
This is the first step in a larger piece of work.

We will use DateTime to store when a user is deactivated, this will
replace the `active` attribute:

- when the deactivated_at is nil the user is active
- when the deactivated_at has a date time the user has been deactivated

We have no way to know when a user was deactivated (hence the work!) so
we'll use the updated_at to be our best guess.

At this point the application is broken, but we'll keep this commit
separate and fix that next.