Update for deactivated_at
We recently dropped the `active` column and want to retain the behaviour
using the `deactivated_at` value:

- if deactivated_at is nil the user is active
- if deactivated_at is present the user has been deactivated

This works creates a `active` method on User and a `active?` alias, to
replicate the old behaviour.

We update the two scopes so they work based on deactivated_at.

Finally, we adjust the specs setup to make sure we are setting a
deactivated_at date not the old active value.

At this point the application is working but the tests will fail and an
attempt to activate or deactivate a user will fail, we'll fix that next.