Allow users to be deactivated and re-activated
This work builds on switching from an `active` boolean to a
`deactivated_at` date and time. Later this switch will allow us to know
how long a user has been deactivated for and 'anonymise' them if they
meet our retention policies.

Here we are enabling users to be deactivated and re-activated in the
same was as they were before, this work is temporary as we plan to
introduce a new UI for managing the state of a user.

We update the service that updates users so it can set or nil the
deactivated_at date and time, we must set it to nil when re-activating.

At this point the application is working and out tests our passing, this
is the smallest work we can do to get out new model into the
application.