Fix: AppicationController spec
Due to some issues between the rpsec-rails gem and Rails 7, possibly
this:

https://github.com/rspec/rspec-rails/issues/2773

this spec stops working in Rails 7.

To work around this, we move in closer and test the behaviour of the
before action itself asserting that the attribute is correct and using
it via the User is also correct.

We no longer need the `reset_all` calls around every spec as we believe
these are being called for us in later version of rspec-rails:

https://github.com/rspec/rspec-rails/pull/2752

We no longer need the DummyController as it was likely not actually
under test.

We take confidence in the other coverage we have that everything is
working as expected for this feature.