Administrative interface for user additional organisations
We now have the ability to add additional organisations from the app itself
via the create/edit user forms.

This replaces the radio list of BEIS + partner organisations with a dropdown,
for the primary organisation, and also adds a checkbox list of partner
organisations for the additional organisations. Appropriate explanatory hint
text is also provided.

We also have a sprinkling of JavaScript to enhance the form such that a user's
additional organisation selections cannot include their primary organisation.
As a safeguard for users who do not have JavaScript enabled, we have validation
in the model for this too.