Fix flaky healthcheck spec
When run on a box with another project that had left unconsumed jobs
in the Sidekiq retry queue, the health check hash matching spec would
fail. In reality we don't care about the contents of the local Sidekiq
Redis DB, so we can just match on the keys we do care about and not
worry about the state of the world with Redis.