Rails 7.0: default disable_to_s_conversion = true
Deprecate passing a format to #to_s in favor of #to_fs in Array, Range,
Date, DateTime, Time, BigDecimal, Float and, Integer.

This deprecation is to allow Rails application to take advantage of a
Ruby 3.1 optimization that makes interpolation of some types of objects
faster.

We do not set custom Date formats in Date::DATE_FORMATS[...] and so this
is unlikely to affect us.