Add an .editorconfig
A lot of editors support https://editorconfig.org out of the box.
Some, like VS Code, require a plugin. But the main benefit of this
little dotfile is that at its most basic it tells your editor to insert
a newline at the end of the file when you've forgotten, thereby
preventing the nasty little red circle of doom on GitHub when you forgot
to add one yourself.