Fix deprecation for instance variables in `render`
Our offences are confined to places where we've used `Activity::Tab`.
This class sends a message to itself in order to set a list of
`@instance_variables` for the given tab. The `ActivitiesController` then
uses `Activity::Tab#locals` to copy a hash of all the instance variables
in the instance of the `Tab` with keys like

`{ :@tab_name => 'financials', :@actuals => [...] }`

to the locals of the template that is about to be rendered. The most
surprising behaviour here is that the `@activity` that is set in the
controller ends up being overwritten by the `@activity` in the
`Tab#locals` method, which is actually an `ActivityPresenter` – a
decorator that shares some identical method names like `total_spend`
which have view-specific behaviours in the Presenter version. We need to
keep defaulting to Presenter.

Make `Tab#locals` pass vanilla keys like `:activity`, and stop using
`@instance_variable`s in app/view/activities/<tab>.html.haml templates.
e.g. `activity.total_spend` instead of `@activity.total_spend`

Also fix broken tabs: Either we can pass locals down assiduously across
a lot of tabs or we can just set `@tab_name` in the ActivitiesController
and let sleeping dogs lie (using `@tab_name=` in tab_nav_item_spec.rb
becomes necessary).

Example deprecation that this commit fixes:

DEPRECATION WARNING: Passing instance variables to `render` is
deprecated.
In Rails 7.1, @activities, @activity, @current_user, @implementing_organisation_presenters, and @tab_name will be ignored.
 (called from block (2 levels) in show at /Users/russ/dev/roda/app/controllers/activities_controller.rb:52)