Functionality to anonymise deactivated users
Here we add functionality to anonymise an already deactivated user.

An anonymised user has:
- a name which is simply "Deleted User <database id>"
- an email address which is "deleted.user.<database id>@<original domain name>"
- no telephone number

In the frontend, an admin user will see an "Anonymise" button when editing
an inactive user; this leads to a confirmation screen which submits a form to
UsersController#anonymise_update, which calls the AnonymiseUser service
object.