Put Level B behind current_user ROLLOUT
We'd like to ship this to a place with real user data to get feedback,
given that access to actual data for dev is not arriving in a timely
manner. We'd also like to restrict that to just one user for now.

Hide the link for them (it's vanishingly unlikely anyone's going to
the URL for these, and they're protected by the ExportPolicy in any
case).

Run

```
ROLLOUT.activate_user(
  :level_b_exports,
  User.find_by!(email: 'whoever.welike@odamanagement.org')
)
```

on prod to enable, and visit

https://report-official-development-assistance.service.gov.uk/rollout/features/level_b_exports

(only set up on prod at present so if you need this for other envs,
don't forget to add a `level_b_exports` feature through the appropriate
env rollout UI)

to set live for everyone later.