Anonymise users background job
This gives us a simple background job which anonymises users who have
been deactivated for more than 5 years. We introduce the sidekiq-scheduler
gem for the purpose of scheduling jobs in a cron-like manner.

An ADR justifies the inclusion of sidekiq-scheduler.

(Also worth noting that the sidekiq_scheduler_spec basically just checks for
typos in the Sidekiq schedule config, in case we ever need to tweak that
again or add additional jobs.)