Level B exports: fix specs for rows
Our factories actually produce a number of activities as side effects
during creates, and this can produce intermittent failures when we
assume there is only one activity for a fund. Ideally, we'd change our
factories and specs so that such side effects are limited, but given
that that would be a Large Job, instead ensure we are looking up our
subject row based on the programme's RODA identifier.

Use the existing `row_for` for this, but fix a subtle bug that meant
it would only work for ISPF files (don't use `FIELDS`, which are both
raw and dynamic, use `export.headers`, which are cooked)