Fix Rails UJS links not working
@rails/ujs is used to transform links to enable deletes. It creates a hidden
form and submits the request as a POST request with a hidden field
_method=delete.

A recent change means this is not working, users therefore cannot perform
important actions.

Calling `start` isn't supposed to be necessary, but seems to fix the issue.

We may need to do a review of these links to see if we can use an improved
method vs Rails UJS.