Allow editing budgets associated with old reports
This has been requested by DSIT colleagues. They've confirmed that
budget information is not directly reported on and is only used for
preparatory work that takes place outside of the service, so editing
budgets associated with approved reports should not cause issues. We
already allow users to provide explanatory comments alongside any
changed to budgetse

We still want the requirement for there to be an editable report for the
associated organisation and fund