Allow editing budgets with no associated report
This is a follow on from 1056fe40b0f95c102cc218c3d05c8d5d9d824146

We found that some budgets aren't associated with a report at all. This
means they fail validation when attempting to update. They should still
be updatable as long as there's a current editable report, so this
associates that report with the budget in such cases