Update GOV.UK Frontend to v5.11
This came about while investigating a report of the "Show" password
component not working. It turns out the "Show" password functionality
was added to the `govuk-form-builder` component in [version 5.3.0](
https://github.com/x-govuk/govuk-form-builder/releases/tag/v5.3.0) but
this wasn't working with our `package.json`'s version 3.11.

Upgrading two major versions feels a bit scary, but I've tweaked styles
where necessary and tested all the JS files and all appears to be
working.

Co-authored-by: Ynda Jas <yndajas@gmail.com>