Update Partner Organisation IDs for specified Activities
We have been provided with a list of Activities where the Partner Organisation
ID requires correction.

For quick turnaround, this was run in the Rails console on prod to avoid going
through the release process; it should not be run again.

Refer to [Zendesk ticket #20965](https://dxw.zendesk.com/agent/tickets/20965)
for full details.