Ignore case when looking up Level C-reporting orgs
in KPI script.

This rake task was failing to produce the expected output on production
as some Organisations have upper case names.

Alternatively, we could resolve this with an ID, but we're unsure
whether identifiers are consistent across environments.

In a perfect world, we'd allow users at DSIT to specify in the UI
whether or not an Organisation reports at Level C or D e.g. with a
checkbox, but this is a larger piece of work.